requirements.txt
numpy
pandas
nltk
Pillow